Adds context to `acctest.PreCheckDirectoryService`
