+++ internal/service/transfer/acc_test.go
@@ -1,0 +1,26 @@
+package transfer_test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/aws/aws-sdk-go/service/transfer"
+	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
+	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+)
+
+func testAccPreCheck(ctx context.Context, t *testing.T) {
+	conn := acctest.Provider.Meta().(*conns.AWSClient).TransferConn()
+
+	input := &transfer.ListServersInput{}
+
+	_, err := conn.ListServersWithContext(ctx, input)
+
+	if acctest.PreCheckSkipError(err) {
+		t.Skipf("skipping acceptance testing: %s", err)
+	}
+
+	if err != nil {
+		t.Fatalf("unexpected PreCheck error: %s", err)
+	}
+}
